缓存适配
